submission woes

+++ Assignment 1/suboptimal3/costs.cpp
@@ -4,7 +4,7 @@ double evalCost() {
 		ans += cc*dashpos[idx1].size();
 		for(int idx2 = idx1+1; idx2 < k ; ++idx2) {
 			for(int pos=0; pos<length; ++pos) {
-				ans += mc[ charToIdx[ sequences[idx1][pos] ] ][ charToIdx[ sequences[idx2][pos] ] ];
+				ans += mc[ charToIdx[ (int) sequences[idx1][pos] ] ][ charToIdx[ (int) sequences[idx2][pos] ] ];
 			}
 		}
 	}
@@ -17,10 +17,10 @@ double evalCost_move(Move move) {
 	//ASSERT : No cost for cc.
 	for(int sidx=0; sidx<k; ++sidx) {
 		if (sidx != move.idx) {
-			double newcost = mc[ charToIdx[ sequences[move.idx][move.newDptr] ] ][ charToIdx[ sequences[sidx][move.origDptr] ] ]
-			 				+ mc[ charToIdx[ sequences[move.idx][move.origDptr] ] ][ charToIdx[ sequences[sidx][move.newDptr] ] ];
-			double oldcost = mc[ charToIdx[ sequences[move.idx][move.origDptr] ] ][ charToIdx[ sequences[sidx][move.origDptr] ] ] 
-							+ mc[ charToIdx[ sequences[move.idx][move.newDptr] ] ][ charToIdx[ sequences[sidx][move.newDptr] ] ];
+			double newcost = mc[ charToIdx[ (int) sequences[move.idx][move.newDptr] ] ][ charToIdx[ (int) sequences[sidx][move.origDptr] ] ]
+			 				+ mc[ charToIdx[ (int) sequences[move.idx][move.origDptr] ] ][ charToIdx[ (int) sequences[sidx][move.newDptr] ] ];
+			double oldcost = mc[ charToIdx[ (int) sequences[move.idx][move.origDptr] ] ][ charToIdx[ (int) sequences[sidx][move.origDptr] ] ] 
+							+ mc[ charToIdx[ (int) sequences[move.idx][move.newDptr] ] ][ charToIdx[ (int) sequences[sidx][move.newDptr] ] ];
 			dcost += (newcost - oldcost);
 		}
 	}
